Update a JAGS model

This function updates a JAGS model created by created by function jags in package jagsUI for a specified number of iterations.

# S3 method for class 'jagsUI'
update(object, parameters.to.save=NULL, n.adapt=NULL, 
  n.iter, n.thin=NULL, modules=c('glm'), factories=NULL, DIC=NULL, 
  codaOnly=FALSE, verbose=TRUE, ...)

Arguments

object

A jagsUI or jagsUIbasic-class object to update.

parameters.to.save

Character vector of the names of the parameters in the model which should be monitored. Defaults to the saved parameter set from the original model run.

n.adapt

Number of iterations to run in the JAGS adaptive phase. The default is NULL, which will result in the function running groups of 100 adaptation iterations (to a max of 10,000) until JAGS reports adaptation is sufficient. If you set n.adapt manually, 1000 is the recommended minimum value.

n.iter

Number of iterations to update for each chain.

n.thin

Thinning rate. Must be a positive integer. Defaults to the thinning rate of the original model run.

modules

List of JAGS modules to load before analysis. By default only module 'glm' is loaded (in addition to 'basemod' and 'bugs'). To force no additional modules to load, set modules=NULL.

factories

Optional character vector of factories to enable or disable, in the format <factory> <type> <setting>. For example, to turn TemperedMix on you would provide 'mix::TemperedMix sampler TRUE' (note spaces between parts). Make sure you have the corresponding modules loaded as well.

DIC

Option to report DIC and the estimated number of parameters (pD). Defaults to the same setting as the original model to updated.

codaOnly

Optional character vector of parameter names for which you do NOT want to calculate detailed statistics. This may be helpful when you have many output parameters (e.g., predicted values) and you want to save time. For these parameters, only the mean value will be calculated but the mcmc output will still be found in $sims.list and $samples.

verbose

If set to FALSE, all text output in the console will be suppressed as the function runs (including most warnings).

...

Further arguments pass to or from other methods.

Author

Ken Kellner contact@kenkellner.com.